backend/pyproject.toml
[tool.poetry]
name = "habitflow"
version = "1.0.0"
description = "HabitFlow - Production-ready habit tracking and gamification platform with virtual forest mechanics"
authors = ["MD Salman Chowdhury (MSCRAFI) <salman@scrafi.dev>"]
readme = "README.md"
homepage = "https://habitflow.scrafi.dev"
repository = "https://github.com/MSCRAFI/habitflow"
keywords = ["habits", "tracking", "gamification", "django", "forest", "productivity"]
classifiers = [
    "Development Status :: 5 - Production/Stable",
    "Environment :: Web Environment",
    "Framework :: Django",
    "Framework :: Django :: 5.0",
    "Intended Audience :: End Users/Desktop",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.11",
    "Topic :: Internet :: WWW/HTTP :: Dynamic Content",
    "Topic :: Office/Business :: Scheduling",
    "Topic :: Games/Entertainment"
]
packages = [
    { include = "backend" }
]

[tool.poetry.dependencies]
python = "^3.11"

# Django & DRF
django = "^5.0.8"
djangorestframework = "^3.15.0"
django-cors-headers = "^4.3.1"
django-filter = "^23.5"

# Authentication & JWT
djangorestframework-simplejwt = "^5.5.1"

# Database & ORM
psycopg2-binary = "^2.9.9"  # PostgreSQL adapter

# API Documentation
drf-spectacular = "^0.27.2"

# Image handling
Pillow = "^10.4.0"

# Async tasks
# Development
django-extensions = "^3.2.3"

[tool.poetry.group.dev.dependencies]
# Code Quality
black = "^23.12.0"
isort = "^5.13.2"
flake8 = "^6.1.0"
pylint = "^3.0.3"
mypy = "^1.7.1"
django-stubs = "^4.2.7"

# Testing
pytest = "^7.4.3"
pytest-django = "^4.7.0"
pytest-cov = "^4.1.0"
factory-boy = "^3.3.0"

# Development Tools
django-debug-toolbar = "^4.2.0"
ipython = "^8.19.0"
pre-commit = "^3.6.0"

[tool.black]
line-length = 100
target-version = ['py311']
include = '\.pyi?$'
extend-exclude = '''
/(
  # directories
  \.eggs
  | \.git
  | \.hg
  | \.mypy_cache
  | \.tox
  | \.venv
  | build
  | dist
  | migrations
)/
'''

[tool.isort]
profile = "black"
line_length = 100
multi_line_mode = 3
include_trailing_comma = true
force_grid_wrap = 0
use_parentheses = true
ensure_newline_before_comments = true
skip = ["migrations", ".venv"]

[tool.mypy]
python_version = "3.11"
check_untyped_defs = true
no_implicit_optional = true
warn_redundant_casts = true
warn_unused_ignores = true
warn_no_return = true
warn_unreachable = true
strict_equality = true
plugins = ["mypy_django_plugin.main"]

[tool.django-stubs]
django_settings_module = "core.settings.local"

[tool.pytest.ini_options]
DJANGO_SETTINGS_MODULE = "core.settings.local"
python_files = ["tests.py", "test_*.py", "*_tests.py"]
testpaths = ["tests"]
addopts = "--cov=. --cov-report=html --cov-report=term-missing"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
